@@ -1,26 +1,30 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
-ruby '2.3.4'
 
-gem 'rails',                      '4.2.10'
+ruby '2.6.5'
 
-# Server, Rack & Middlewares
+gem 'brcobranca'
+gem 'parseline'
 gem 'puma'
-gem 'rack-canonical-host'
-
-gem "rghost"
-gem "rghost_barcode"
-gem "parseline"
-gem "brcobranca", git: 'https://github.com/kivanio/brcobranca'
+gem 'rails'
+gem 'rghost'
+gem 'rghost_barcode'
+gem 'sprockets', '3.7.2'
 
 gem 'sass-rails'
 gem 'uglifier'
+# Reduces boot times through caching; required in config/boot.rb
+gem 'bootsnap', '>= 1.4.2', require: false
 
-group :development, :test do
-  gem 'byebug'
-  gem 'spring'
+group :development do
   gem 'foreman'
+  gem 'rubocop', '0.81.0', require: false
+  gem 'rubocop-performance', '1.5.2', require: false
+  gem 'rubocop-rails', '2.5.2', require: false
+  gem 'spring'
+  gem 'spring-watcher-listen', '~> 2.0.0'
 end
-
 group :production do
-  gem 'rails_12factor'
+  gem 'rack-canonical-host'
 end
