@@ -1,8 +1,10 @@
-require File.expand_path('../boot', __FILE__)
+# frozen_string_literal: true
 
-require "action_controller/railtie"
-require "action_view/railtie"
-require "sprockets/railtie"
+require_relative 'boot'
+
+require 'action_controller/railtie'
+require 'action_view/railtie'
+require 'sprockets/railtie'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -10,18 +12,11 @@ Bundler.require(*Rails.groups)
 
 module BrcobrancaExemplo
   class Application < Rails::Application
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 6.0
 
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
     config.time_zone = ENV.fetch('TZ', 'Brasilia')
 
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
     config.i18n.enforce_available_locales = false
     config.i18n.default_locale = :"pt-BR"
     config.i18n.locale = :"pt-BR"
