@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 # Load the Rails application.
-require File.expand_path('../application', __FILE__)
+require_relative 'application'
 
 # Initialize the Rails application.
 Rails.application.initialize!
