@@ -1,9 +1,8 @@
-workers Integer(ENV['WEB_CONCURRENCY'] || 1)
-threads_count = Integer(ENV['MAX_THREADS'] || 1)
-threads threads_count, threads_count
+max_threads_count = ENV.fetch("RAILS_MAX_THREADS") { 5 }
+min_threads_count = ENV.fetch("RAILS_MIN_THREADS") { max_threads_count }
+threads min_threads_count, max_threads_count
+port        ENV.fetch("PORT") { 3000 }
+environment ENV.fetch("RAILS_ENV") { "development" }
+workers ENV.fetch("WEB_CONCURRENCY") { 2 }
 
 preload_app!
-
-rackup      DefaultRackup
-port        ENV['PORT'] || 3000
-environment ENV['RACK_ENV'] || 'development'
